verilog-ppa-tuning/requirements.txt
torch>=2.5.1
transformers>=4.25.1
datasets>=2.7.1
peft>=0.3.0
accelerate>=0.16.0
numpy>=1.21.0
tqdm>=4.62.0
pyyaml>=6.0
colorlog>=6.7.0
scikit-learn>=1.0.2
tensorboard>=2.10.0
sentencepiece>=0.1.97
jsonlines>=3.1.0
pandas>=1.3.5
matplotlib>=3.5.1
bitsandbytes>=0.37.0  # 用于8bit量化
wandb>=0.13.7  # 可选，用于实验跟踪